refactor(step2): 패키지 경로 수정
