+++ README.md
@@ -34,8 +34,10 @@ cucumber가 제공하는 기능을 적절히 활용해보세요.
 - [x] 중복되는 사전조건(지하철 역들/노선들을 생성한다)은 background를 활용해도 좋을 것 같다.
   - Background 설정
   - AcceptanceContext에서 station객체와 line 객체를 각각 관리 
-- [ ] 패키지 구조의 게층이 잘 그려지지 않는 문제. subway 하위에 도메인으로 나뉠지, 계층별로 나뉠지 통일이 필요해 보임
+- [x] 패키지 구조의 게층이 잘 그려지지 않는 문제. subway 하위에 도메인으로 나뉠지, 계층별로 나뉠지 통일이 필요해 보임
+  - `section`과 `path`패키지를 `subway`하위로 이동
 - [ ] PathFinder 객체의 이름이 PathFinder 보다는 Path나 SubwayMap과 같이 바라볼 수 있을 것 같음. 고민 필요
+  - Path vs PathFinder / 명확한 답을 내리기 어렵다
 - [ ] 검증 시 resopnse.jsonPath().getList("...")와 같은 부분의 각 단계를 메서드로 분리하면 재사용도 가능하고 시나리오가 더욱 명확해질 수 있다.
 - [ ] 거리의 타입이 int 대신 long 인 이유는?
   - 표현할 수 있는 값의 범위가 큰 것이 무조건적인 장점일까? 저장할 데이터의 최소/최대 범위를 고려해서 자료형을 정해야 한다.
