+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -1,8 +1,8 @@
 package nextstep.subway.line.domain;
 
 
-import nextstep.subway.line.section.domain.Section;
-import nextstep.subway.line.section.domain.Sections;
+import nextstep.subway.section.domain.Section;
+import nextstep.subway.section.domain.Sections;
 import nextstep.subway.station.domain.Station;
 
 import javax.persistence.*;
