+++ src/main/java/nextstep/subway/line/ui/LineController.java
@@ -5,10 +5,8 @@ import nextstep.subway.line.application.dto.LineRequest;
 import nextstep.subway.line.application.dto.LineResponse;
 import nextstep.subway.line.application.dto.LineSectionResponse;
 import nextstep.subway.line.application.dto.UpdateLineRequest;
-import nextstep.subway.line.path.PathResponse;
-import nextstep.subway.line.path.PathType;
-import nextstep.subway.line.section.dto.SectionRequest;
-import nextstep.subway.line.section.dto.SectionResponse;
+import nextstep.subway.section.dto.SectionRequest;
+import nextstep.subway.section.dto.SectionResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -66,10 +64,5 @@ public class LineController {
     public ResponseEntity<Void> deleteLineSection(@PathVariable Long id, @RequestParam Long stationId) {
         lineService.deleteSection(id, stationId);
         return ResponseEntity.noContent().build();
-    }
-
-    @GetMapping("/paths")
-    public ResponseEntity<PathResponse> shortestPath(@RequestParam("source") Long source, @RequestParam("target") Long target, @RequestParam("type") PathType type) {
-        return ResponseEntity.ok().body(lineService.getShortestPath(source, target, type));
     }
 }
