+++ src/main/java/nextstep/subway/path/application/PathService.java
@@ -1,0 +1,33 @@
+package nextstep.subway.path.application;
+
+import nextstep.subway.Exception.ErrorCode;
+import nextstep.subway.Exception.SubwayException;
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.line.domain.LineRepository;
+import nextstep.subway.path.domain.Path;
+import nextstep.subway.path.domain.PathType;
+import nextstep.subway.path.application.dto.PathResponse;
+import nextstep.subway.station.domain.Station;
+import nextstep.subway.station.domain.StationRepository;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class PathService {
+    private final StationRepository stationRepository;
+    private final LineRepository lineRepository;
+
+    public PathService(StationRepository stationRepository, LineRepository lineRepository) {
+        this.stationRepository = stationRepository;
+        this.lineRepository = lineRepository;
+    }
+
+    public PathResponse getShortestPath(Long source, Long target, PathType type) {
+        Station sourceStation = stationRepository.findById(source).orElseThrow(() -> new SubwayException(ErrorCode.STATION_NOT_FOUND, ""));
+        Station targetStation = stationRepository.findById(target).orElseThrow(() -> new SubwayException(ErrorCode.STATION_NOT_FOUND, ""));
+        List<Line> lines = lineRepository.findAll();
+
+        return new Path(lines).shortestPath(sourceStation, targetStation, type);
+    }
+}
