+++ src/main/java/nextstep/subway/section/domain/Sections.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.section.domain;
+package nextstep.subway.section.domain;
 
 
 import nextstep.subway.Exception.ErrorCode;
