+++ src/main/java/nextstep/subway/section/dto/SectionRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.section.dto;
+package nextstep.subway.section.dto;
 
 public class SectionRequest {
     private Long upStationId;
