+++ src/test/java/nextstep/cucumber/steps/StationStepDef.java
@@ -8,9 +8,9 @@ import io.restassured.response.Response;
 import nextstep.subway.line.application.dto.LineRequest;
 import nextstep.subway.line.application.dto.LineResponse;
 import nextstep.subway.line.domain.Line;
-import nextstep.subway.line.path.PathType;
-import nextstep.subway.line.section.domain.Section;
-import nextstep.subway.line.section.dto.SectionRequest;
+import nextstep.subway.path.domain.PathType;
+import nextstep.subway.section.domain.Section;
+import nextstep.subway.section.dto.SectionRequest;
 import nextstep.subway.station.application.dto.StationResponse;
 import nextstep.subway.station.domain.Station;
 import nextstep.subway.utils.AcceptanceContext;
