+++ src/test/java/nextstep/subway/acceptance/PathAcceptanceTest.java
@@ -3,8 +3,8 @@ package nextstep.subway.acceptance;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.line.application.dto.LineRequest;
-import nextstep.subway.line.path.PathType;
-import nextstep.subway.line.section.dto.SectionRequest;
+import nextstep.subway.path.domain.PathType;
+import nextstep.subway.section.dto.SectionRequest;
 import nextstep.subway.utils.AcceptanceTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
