+++ src/test/java/nextstep/subway/unit/PathFinderTest.java
@@ -2,10 +2,10 @@ package nextstep.subway.unit;
 
 import nextstep.subway.Exception.SubwayException;
 import nextstep.subway.line.domain.Line;
-import nextstep.subway.line.path.PathResponse;
-import nextstep.subway.line.path.PathFinder;
-import nextstep.subway.line.path.PathType;
-import nextstep.subway.line.section.domain.Section;
+import nextstep.subway.path.application.dto.PathResponse;
+import nextstep.subway.path.domain.Path;
+import nextstep.subway.path.domain.PathType;
+import nextstep.subway.section.domain.Section;
 import nextstep.subway.station.domain.Station;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -37,7 +37,7 @@ public class PathFinderTest {
     @DisplayName("최단 시간 경로 조회")
     @Test
     void shortestDistancePath() {
-        PathResponse pathResponse = new PathFinder(LINES).shortestPath(교대역, 양재역, PathType.DURATION);
+        PathResponse pathResponse = new Path(LINES).shortestPath(교대역, 양재역, PathType.DURATION);
 
         assertThat(pathResponse.getStations()).containsExactly(교대역, 강남역, 양재역);
         assertThat(pathResponse.getDistance()).isEqualTo(24L);
@@ -47,7 +47,7 @@ public class PathFinderTest {
     @DisplayName("최단 거리 경로 조회")
     @Test
     void shortestDurationPath() {
-        PathResponse pathResponse = new PathFinder(LINES).shortestPath(교대역, 양재역, PathType.DISTANCE);
+        PathResponse pathResponse = new Path(LINES).shortestPath(교대역, 양재역, PathType.DISTANCE);
 
         assertThat(pathResponse.getStations()).containsExactly(교대역, 남부터미널역, 양재역);
         assertThat(pathResponse.getDistance()).isEqualTo(23L);
@@ -57,7 +57,7 @@ public class PathFinderTest {
     @DisplayName("에러_최단 거리 경로 조회_출발역 도착역 같음")
     @Test
     void error_shortestPath_target_source_same() {
-        assertThatThrownBy(() -> new PathFinder(LINES).shortestPath(교대역, 교대역, PathType.DISTANCE))
+        assertThatThrownBy(() -> new Path(LINES).shortestPath(교대역, 교대역, PathType.DISTANCE))
                 .isInstanceOf(SubwayException.class)
                 .hasMessage("출발역과 도착역이 같습니다.");
     }
@@ -65,7 +65,7 @@ public class PathFinderTest {
     @DisplayName("에러_최단 거리 경로 조회_출발역 도착역 연결되지 않음")
     @Test
     void error_shortestPath_target_source_not_connected() {
-        assertThatThrownBy(() -> new PathFinder(LINES).shortestPath(교대역, 미금역, PathType.DISTANCE))
+        assertThatThrownBy(() -> new Path(LINES).shortestPath(교대역, 미금역, PathType.DISTANCE))
                 .isInstanceOf(SubwayException.class)
                 .hasMessage("연결되지 않은 역 정보입니다.");
     }
